+++ spring-cloud-starter-tencent-polaris-ratelimit/src/main/java/com/tencent/cloud/polaris/ratelimit/filter/QuotaCheckReactiveFilter.java
@@ -23,6 +23,7 @@ import java.time.Duration;
 import java.util.Objects;
 import java.util.Set;
 
+import com.tencent.cloud.common.constant.HeaderConstant;
 import com.tencent.cloud.common.metadata.MetadataContext;
 import com.tencent.cloud.polaris.ratelimit.config.PolarisRateLimitProperties;
 import com.tencent.cloud.polaris.ratelimit.constant.RateLimitConstant;
@@ -30,6 +31,7 @@ import com.tencent.cloud.polaris.ratelimit.resolver.RateLimitRuleArgumentReactiv
 import com.tencent.cloud.polaris.ratelimit.spi.PolarisRateLimiterLimitedFallback;
 import com.tencent.cloud.polaris.ratelimit.utils.QuotaCheckUtils;
 import com.tencent.cloud.polaris.ratelimit.utils.RateLimitUtils;
+import com.tencent.polaris.api.pojo.RetStatus;
 import com.tencent.polaris.ratelimit.api.core.LimitAPI;
 import com.tencent.polaris.ratelimit.api.rpc.Argument;
 import com.tencent.polaris.ratelimit.api.rpc.QuotaResponse;
@@ -114,6 +116,10 @@ public class QuotaCheckReactiveFilter implements WebFilter, Ordered {
 					response.getHeaders().setContentType(MediaType.TEXT_HTML);
 					dataBuffer = response.bufferFactory().allocateBuffer()
 							.write(rejectTips.getBytes(StandardCharsets.UTF_8));
+				}
+				response.getHeaders().add(HeaderConstant.INTERNAL_CALLEE_RET_STATUS, RetStatus.RetFlowControl.getDesc());
+				if (Objects.nonNull(quotaResponse.getActiveRule())) {
+					response.getHeaders().add(HeaderConstant.INTERNAL_ACTIVE_RULE_NAME, quotaResponse.getActiveRule().getName().getValue());
 				}
 				return response.writeWith(Mono.just(dataBuffer));
 			}